move loadout-related buttons
